orm-spring-1.0.1# 进一步提取为spring框架版

+++ quickjava-orm/src/main/java/org/quickjava/orm/wrapper/AbstractModelWrapper.java
@@ -315,7 +315,7 @@ public abstract class AbstractModelWrapper<Children extends AbstractModelWrapper
         return model().insert(data);
     }
 
-    public int delete() {
+    public Long delete() {
         return model().delete();
     }
